% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gmm_pp.R
\name{fit_gmm_pp}
\alias{fit_gmm_pp}
\title{Title}
\usage{
fit_gmm_pp(D, K)
}
\arguments{
\item{D}{Dataframe containing crimes to be used to train the model. Should be in format provided by the fix_date_column( function).}

\item{K}{Number of components the Gaussian mixture should use.}
}
\value{
A list containing the fitted rate function of the form \code{f(x,y,t_delta)}, the list of parameters that are used in the rate function, and the list parameters used to scale the data before fitting, so that they may be used on new data.
}
\description{
Title
}
