% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_new_points.R
\name{generate_new_points}
\alias{generate_new_points}
\title{Generate new locations of crimes based on the fitted model}
\usage{
generate_new_points(
  gmm_pp,
  t_delta = 1,
  area_bounds = list(c(-87.84918, 41.6608), c(-87.40973, 41.98827))
)
}
\arguments{
\item{gmm_pp}{A fitted poisson process returned from the fit_gmm_pp function.}

\item{t_delta}{The length of time in days that new points should be generated for. Crimes can be generated starting immediately after the last crime that was trained on.}

\item{area_bounds}{A list of two vertices that specify the box in which points are generated. The vertices are given in the form \code{c(Longitude, Latitude)}
The default is the area shown in the mapping functions, i.e. \code{list(c(-87.84918,41.6608), c(-87.40973,41.98827))}.}
}
\value{
A dataframe containing Longitude, Latitude, and Date columns for each generated crime. This is enough to be passed to the \code{map_crime_points} and \code{map_crime_kde} functions.
}
\description{
Generate new locations of crimes based on the fitted model
}
