% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_mixture_fitting.R
\name{gmm_em_fit}
\alias{gmm_em_fit}
\title{Gaussian Mixture Fitting via the EM Algorithm}
\usage{
gmm_em_fit(X, K, mus = NULL)
}
\arguments{
\item{X}{Data matrix with observations as rows.}

\item{K}{Number of components the Gaussian mixture should use}

\item{mus}{Optional parameter that provides starting centres for the Gaussian components. Matrix whose rows are vectors giving the position of each centre. If not provided, centres will be chosen at random from a normal distribution based on the entire data matrix.}
}
\value{
A list containing the fitted parameters \code{ws}, \code{mus}, and \code{Sigmas}. \code{ws} is a vector containing the weights for each components of the mixture. \code{mus} is a matrix whose rows are the centres of each component of the mixture. \code{Sigmas} is a list containing the covariances matrices for each component of the mixture.
}
\description{
Gaussian Mixture Fitting via the EM Algorithm
}
\examples{
X = rbind(c(0,0),c(-1,0),c(1,0),c(0,1),c(0,-1),c(5,5),c(4,5),c(5,6),c(6,5),c(5,4))
K = 2
start_centres = rbind(c(-1,-1),c(6,6))
par_list = chicagomodel:::gmm_em_fit(X,K,start_centres)
}
\keyword{internal}
