% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_crime_data.R
\name{load_crime_data}
\alias{load_crime_data}
\title{Load Chicago Crime Dataset and Subset by Time/Date and Crime Type}
\usage{
load_crime_data(
  chicago_data_path,
  start_time = NULL,
  end_time = NULL,
  crime = NULL
)
}
\arguments{
\item{chicago_data_path}{String specifying the path of the Chicago crime dataset CSV as available from https://data.cityofchicago.org/Public-Safety/Crimes-2001-to-Present/ijzp-q8t2}

\item{start_time}{A start datetime of the form "yyyy-mm-dd hh:mm:ss" where "hh" is 24 hour format, local time with daylight savings.}

\item{end_time}{An end datetime of the same form as \code{start_time}. Note that if only a date is provided, time will be set to 00:00:00, meaning the day specified in \code{end_time} will not be included.}

\item{crime}{Either a string or a list of strings each specifying a crime type as stored in the \code{Primary.Type} column of the Chicago crime dataset. The possible crimes are DECEPTIVE PRACTICE, THEFT,
CRIMINAL DAMAGE, ASSAULT, BATTERY, OTHER OFFENSE, NARCOTICS, WEAPONS VIOLATION, MOTOR VEHICLE THEFT, CRIMINAL TRESPASS, BURGLARY, INTERFERENCE WITH PUBLIC OFFICER, ARSON, ROBBERY, CRIM SEXUAL ASSAULT,
CRIMINAL SEXUAL ASSAULT, PROSTITUTION, PUBLIC PEACE VIOLATION, OFFENSE INVOLVING CHILDREN, LIQUOR LAW VIOLATION, CONCEALED CARRY LICENSE VIOLATION, SEX OFFENSE, STALKING, GAMBLING, HOMICIDE, OBSCENITY,
INTIMIDATION, KIDNAPPING, HUMAN TRAFFICKING, NON-CRIMINAL, OTHER NARCOTIC VIOLATION, PUBLIC INDECENCY.}
}
\value{
A dataframe containing the crime data between the specified times / dates and of the specified types(s), where the datetimes have been converted to POSIXct format. The data is cleaned of any entries where location data is missing or set as a default value.
}
\description{
Load Chicago Crime Dataset and Subset by Time/Date and Crime Type
}
