% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{map_crime_kde}
\alias{map_crime_kde}
\title{Visualise Kernel Density Estimate of Crime on a Map}
\usage{
map_crime_kde(data)
}
\arguments{
\item{data}{The crime dataframe, most likely fetched using the \code{load_crime_data} function.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualise Kernel Density Estimate of Crime on a Map
}
