% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{map_gmm_pp}
\alias{map_gmm_pp}
\title{Visualise Fitted Gaussian Mixture Poisson Process Model}
\usage{
map_gmm_pp(data, gmm_pp)
}
\arguments{
\item{data}{The crime dataframe, most likely fetched using the \code{load_crime_data} function.}

\item{gmm_pp}{The list outputted by \code{fit_gmm_pp}, containing the necessary information and parameters for the model.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualise Fitted Gaussian Mixture Poisson Process Model
}
