% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_number_of_crimes.R
\name{predict_number_of_crimes}
\alias{predict_number_of_crimes}
\title{Predict the number of crimes over a certain rectangular area and time period}
\usage{
predict_number_of_crimes(gmm_pp, lower, upper, t_delta)
}
\arguments{
\item{gmm_pp}{A fitted model given by the \code{fit_gmm_pp} function}

\item{lower}{The lower left vertex of the region as a vector, i.e. \code{c(x1,y1)}}

\item{upper}{The upper right vertex of the region as a vector, i.e. \code{c(x2, y2)}}

\item{time_delta}{How long in days the period to predict should be. For example, if you trained on February and wanted to predict the number of crimes in March,}
}
\value{
The predicted number of crimes.
}
\description{
Predict the number of crimes over a certain rectangular area and time period
}
